refactor(api): move service functions to file
