+++ src/app/api/users/route.ts
@@ -2,43 +2,12 @@ import crypto from 'crypto';
 import { Prisma } from '@prisma/client';
 import { bcrypt } from 'hash-wasm';
 import { z } from 'zod';
-import prisma from '@/lib/prisma';
-
-export const userBodySchema = z.object({
-  email: z.string().email(),
-  name: z.string().nullish(),
-  password: z.string().min(8),
-});
-
-export const HASH_ROUNDS = 10;
-
-export const getAllUsers = async () => {
-  return prisma.user.findMany({
-    select: {
-      id: true,
-      email: true,
-      name: true,
-      createdAt: true,
-      updatedAt: true,
-    },
-  });
-};
-
-export const createUser = async (data: Prisma.UserCreateInput) => {
-  return prisma.user.create({
-    data,
-    select: {
-      id: true,
-      email: true,
-      name: true,
-      createdAt: true,
-      updatedAt: true,
-    },
-  });
-};
-
-export type GetAllUsers = Prisma.PromiseReturnType<typeof getAllUsers>;
-export type CreateUser = Prisma.PromiseReturnType<typeof createUser>;
+import {
+  HASH_ROUNDS,
+  createUser,
+  getAllUsers,
+  userBodySchema,
+} from '@/app/api/users/service';
 
 const GET = async () => {
   const result = await getAllUsers();
