+++ src/components/forms/SignUpForm.tsx
@@ -14,7 +14,7 @@ import axios, { AxiosError } from 'axios';
 import { useRouter } from 'next/navigation';
 import { useForm } from 'react-hook-form';
 import { z } from 'zod';
-import { type CreateUser, userBodySchema } from '@/app/api/users/route';
+import { type CreateUser, userBodySchema } from '@/app/api/users/service';
 
 export const signUpSchema = userBodySchema
   .extend({
