+++ src/components/layouts/Navbar.tsx
@@ -7,7 +7,7 @@ import axios from 'axios';
 import Link from 'next/link';
 import { useParams } from 'next/navigation';
 import { useSession } from 'next-auth/react';
-import { type GetChat } from '@/app/api/users/[userId]/chat/route';
+import { type GetChat } from '@/app/api/users/[userId]/chat/service';
 import { ChatMenu } from '@/components/modules/ChatMenu';
 import { ProfileMenu } from '@/components/modules/ProfileMenu';
